requirements.txt
catalog-config==0.0.2
catalog-utility-kit==1.0.3
certifi==2021.10.8
charset-normalizer==2.0.9
elastic-transport==8.1.2
elasticsearch==7.16.2
idna==3.3
psycopg2-binary==2.9.2
requests==2.26.0
urllib3==1.26.7
